Add CKAN 2.0 templates for Google Analytics tracking
Everything should still work as it did with CKAN 1.x templates _if_
ckan.legacy_templates is set in your ini file.

The event tracking javascript (ckanext-googleanalytics.js) still needs
to be updated for CKAN 2.0.

Also there is some code in plugin.py:filter() that adds something to the
package and group pages that I haven't implemented for CKAN 2.0 yet.

Tests need to be updated.

README needs to be updated.