Attempts to provide support for v3 of the Google Analytics API.
Since version 2 of the Google Analytics API was shelved the download
of previous data was broken.  This commit should, once the oauth dance
has been complete, allow it to work again by changing all of the calls.
See the README for instructions on setting up OAUTH.  Other bug fixes
are:

- Resources views are now clearly identified
- Package Ids are set for /dataset/edit requests
- Updated README