Remove ckan.authz
It has been removed in CKAN 2.0, trying to import it makes CKAN crash on
startup.

This means that get_top_packages() is broken, and needs to be rewritten
to work with CKAN 2.0. At least by removing the import the plugin can
be loaded and its other functions used but if you try to use
get_top_packages() it will crash.