Use g not config in templates
config is no longer available to templates (changed in CKAN core), use g
instead.

This means I have to rename googleanalytics.resource_prefix to
googleanalytics_resource_prefix, because it's now an attribute and can't
have a . in it.

Should probably be using g rather than config to access the setting in
the Python code as well..